+++ mango-pileup/bdgenomics/mango/js/src/test/index.ts
@@ -1,4 +1,4 @@
 // import all tests here, otherwise if we include them in karma.conf.js it will all be separate bundles
 import './widget-utils';
-import './utils-test';
-import './pileupViewer-test';
+import './utils';
+import './pileupViewer';
